# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Odontología",
    "description": "API para el sistema de gestión de pacientes, tratamientos y citas odontológicas.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000"
    }
  ],
...
  "paths": {
    "/usuarios/login": {
      "post": {
        "tags":["Usuarios"],
        "summary": "Iniciar sesión",
        "description": "Este endpoint permite iniciar sesión en el sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["correo", "contrasena"],
                "properties": {
                  "correo": { "type": "string", "format": "email" },
                  "contrasena": { "type": "string" }
                },
                "example": {
                  "correo": "usuario@example.com",
                  "contrasena": "miContrasena"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mensaje": { "type": "string" },
                    "usuario": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "nombre": { "type": "string" }
                      }
                    }
                  },
                  "example": {
                    "mensaje": "Login exitoso",
                    "usuario": { "id": 1, "nombre": "Carlos" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales incorrectas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "mensaje": { "type": "string" } }
                },
                "example": { "mensaje": "Credenciales incorrectas" }
              }
            }
          }
        }
      }
    },
...
    "/usuarios/registroPaciente": {
      "post": {
        "tags": ["Usuarios"],
        "summary": "Registrar un nuevo paciente",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nombre", "apellido", "cedula", "telefono", "correo", "direccion", "fecha_registro", "contrasena"],
                "properties": {
                  "nombre": { "type": "string" },
                  "apellido": { "type": "string" },
                  "cedula": { "type": "string" },
                  "telefono": { "type": "string" },
                  "correo": { "type": "string", "format": "email" },
                  "direccion": { "type": "string" },
                  "fecha_registro": { "type": "string", "format": "date" },
                  "contrasena": { "type": "string" }
                },
                "example": {
                  "nombre": "Ana",
                  "apellido": "Gómez",
                  "cedula": "987654321",
                  "telefono": "3001234567",
                  "correo": "ana@example.com",
                  "direccion": "Calle 123",
                  "fecha_registro": "2025-06-09",
                  "contrasena": "contraseña123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paciente registrado con contraseña",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "mensaje": { "type": "string" } },
                  "example": { "mensaje": "Paciente registrado con contraseña" }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/editarPaciente/{id}": {
      "put": {
        "tags": ["Usuarios"],
        "summary": "Editar datos de un paciente",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "apellido": { "type": "string" },
                  "telefono": { "type": "string" },
                  "correo": { "type": "string", "format": "email" },
                  "direccion": { "type": "string" }
                },
                "example": {
                  "nombre": "Ana",
                  "apellido": "Gómez",
                  "telefono": "3000000000",
                  "correo": "ana2@example.com",
                  "direccion": "Nueva dirección"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paciente actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "mensaje": { "type": "string" } },
                  "example": { "mensaje": "Paciente actualizado" }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/eliminarPaciente/{id}": {
      "delete": {
        "tags": ["Usuarios"],
        "summary": "Eliminar un paciente",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Paciente eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "mensaje": { "type": "string" } },
                  "example": { "mensaje": "Paciente eliminado" }
                }
              }
            }
          }
        }
      }
    },
...
  }
}